+++ maestro-ios-driver/src/main/kotlin/xcuitest/XCTestDriverClient.kt
@@ -17,7 +17,7 @@ import java.net.ConnectException
 import java.net.SocketTimeoutException
 import java.net.UnknownHostException
 import java.util.concurrent.TimeUnit
-import xcuitest.process.XCTestStarter
+import xcuitest.starter.XCTestStarter
 
 const val UI_TEST_RUNNER_APP_BUNDLE_ID = "dev.mobile.maestro-driver-iosUITests.xctrunner"
